IA-2774: reorganize files
