+++ hat/assets/js/apps/Iaso/domains/payments/LotsPayments.tsx
@@ -7,7 +7,7 @@ import { baseUrls } from '../../constants/urls';
 import { PotentialPaymentParams } from './types';
 import { useGetPaymentLots } from './hooks/requests/useGetPaymentLots';
 import { usePaymentLotsColumns } from './hooks/config/usePaymentLotsColumns';
-import { PaymentLotsFilters } from './components/PaymentLotsFilters';
+import { PaymentLotsFilters } from './components/CreatePaymentLot/PaymentLotsFilters';
 import { SimpleTableWithDeepLink } from '../../components/tables/SimpleTableWithDeepLink';
 
 type Props = {
