+++ hat/assets/js/apps/Iaso/domains/payments/PotentialPayments.tsx
@@ -13,11 +13,11 @@ import { PotentialPaymentParams, PotentialPayment } from './types';
 import { useGetPotentialPayments } from './hooks/requests/useGetPotentialPayments';
 import { TableWithDeepLink } from '../../components/tables/TableWithDeepLink';
 import { baseUrls } from '../../constants/urls';
-import { PotentialPaymentsFilters } from './components/PotentialPaymentsFilters';
+import { PotentialPaymentsFilters } from './components/CreatePaymentLot/PotentialPaymentsFilters';
 import { redirectTo } from '../../routing/actions';
 import { usePaymentColumns } from './hooks/config/usePaymentColumns';
 import { Selection } from '../orgUnits/types/selection';
-import { AddPaymentLotDialog } from './components/PaymentLotDialog';
+import { AddPaymentLotDialog } from './components/CreatePaymentLot/PaymentLotDialog';
 
 type Props = {
     params: PotentialPaymentParams;
